+++ 01/allocator.h
@@ -1,8 +1,8 @@
 #ifndef	_ALLOCATOR_H_
 #define	_ALLOCATOR_H_
 
-void makeAllocator(uint32_t);
-char * alloc(uint32_t);
+void makeAllocator(size_t);
+char * alloc(size_t);
 void reset();
 
 #endif
